entities fix int to long

+++ TMS.Api/TMS.Api/Models/EventType.cs
@@ -5,7 +5,7 @@ namespace TMS.Api.Models;
 
 public partial class EventType
 {
-    public int EventTypeId { get; set; }
+    public long EventTypeId { get; set; }
 
     public string EventTypeName { get; set; } = null!;
